fixed sparsity calculations; added a function to infer masks for a pruned model
sparsity calculation was bugged: 1. didn't return a float. 2. double counted embedding params, now fixed.
added a manual sparsity calculator that counts zero weights instead of zeros in the mask
added a function that allows for inference of masks for a pruned model. stores the inferred mask in the self.masks variable. checks param.data for zeros, and sets all corresponding entries in the mask to zero.